requirement.txt
altgraph==0.17.4
importlib-metadata==6.8.0
macholib==1.16.3
packaging==23.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyYAML==6.0.1
zipp==3.17.0
